partially fixed issue with updating line number width on new font
There's currently a bug that only affects calls to the editor setFont
function after the first one. The first call will set the correct line
number margin width, but subsequent calls will set the width based on
the previous font value instead of the current one. This may be due to a
delay in the underlying scintilla library receiving the new font info.